Thesis: P(NN(i)=h, (NN(i')=h') back up
This code creates two samples of size 30 each from uniform distribution
and commutes the distance between every two points using  euclidean
distance as a distance metric defined over a torus. Further, using an
unbiased estimator, the probability of the number nearest neighbors of
a point i equalling h and the number of nearest neighbors of a point i’
equalling h’ is estimated.