Test Statistic  uniform data
This code generates two samples from uniform distribution of 3
dimension data as an example but can be used for higher dimensions as
well. It finds the k-NN using euclidean distance defined on a torus as
the distance metric. The test is a weighted proportion of the count of
points with nearest neighbors from the same sample. I use weights
inversely proportional to hypergeometric pmf.